+++ packages/client-react-streaming/src/ManualDataTransport/serialization.test.ts
@@ -1,5 +1,5 @@
 import test, { describe } from "node:test";
-import { revive, stringify } from "./dataTransport.js";
+import { revive, stringify } from "./serialization.js";
 import { outsideOf } from "../util/runInConditions.js";
 import { htmlEscapeJsonString } from "./htmlescape.js";
 import assert from "node:assert";
